feat: expandable component
We have 3 expandable sections in Podman Desktop: one for the learning center and
two in the Kubernetes dashboard. They use two different animation speeds and all
three had the same minor bugs:
 - the text shouldn't move right/left when opening/closing the control
 - the animation is smooth - and then jumps the last bit

The first bug was due to not setting the width of the toggle, the second was
due to an outer div that isn't animated, so it just appears/disappears.

Instead of copying the fixes 3 times this creates an expandable component with
the basic function of these instances.

Notes:
- Moved to ui package, both because extensions may want to use the same control
  and because I expect we'll have a Dashboard component there
- expanded+initialized could be a single property, but since most users will just
  want a simple boolean expanded I kept them separate.

Fixes #.
Part of updating bootc dashboard issue 976.

Signed-off-by: Tim deBoer <git@tdeboer.ca>